Goblins for state transition systems.
Goblins are structured mutators for a type; structured in that the nature of
that mutation is goverened by their genome. GHC Generics provide a way to map
this genome to a Goblin.

The intention is to try to 'breed' goblins for a state transition system, such
that a goblin is bred to cause specific types of failure in that system.

Add goblin instance for Data.Ratio

Add Goblin for `Maybe`

Allow `Gen` actions for env and state.

Single generator for full context.

This is likely not what we want eventually, which will be to use `ProgressiveGen`.
But this will work for the time-being for testing purposes.

Fix up goblins!

Add Goblin explainer.

Appease the formatting tyrant.

Move goblins to a separate library.