Add `cardano-prelude` as a submodule
This is an alternative to `stack.yaml` commit pinning.  I've found
it to be smoother for development. The pain point this addresses is
making and testing local changes (e.g. changes in `cardano-prelude`
which are used/tested from `cardano-chain`) - with the `stack.yaml`
-pinning style, changes must be pushed to a remote or done through
an include of a local git repo. The submodules style lets you make
local modifications, test that they work, then commit and push to
the submodule remote (`cardano-prelude` in this case) before bumping
the submodule tracking hash from the parent (`cardano-chain` in this
case), committing that, and pushing it to the parent's remote.