Remove use of LovelacePortion in genesis generation
Change uses of LovelacePortion to Rational.

The GenesisInitializer was using LovelacePortion for calculations for
what portion of all the money goes to the "richmen" vs everyone else.
This is unnecessary. We can simply use Rational, and use a function
for scaling a Lovelace by a Rational.

LovelacePortion is a (fixed base) representation of a rational in the
range [0..1] that we inherited for Byron compatability for the
thresholds for (mostly unused) update system protocol paramaters.

We should avoid using it for anything else.