Replace only real use of lovelacePortionToDouble
Use lovelacePortionToRational instead. It was being used in applying an
update system threshold paramater to the number of genesis keys to
determine the threshold required.

Since this was being used with the number of genesis keys, rather than
total system stake, then the use of Double was not in practice a
problem, but it's much clearer to stick to our rule that we avoid
floating point completely within the ledger and consensus rules.