Change Lovelace internal rep from Word64 to Natural
We will be able to eliminate a lot of the complexity involved in
manipulating Lovelace values by making most of the operations total,
where currently they are partial.

The first step is changing the internal representation, without changing
anything else. So it still does the overflow checks at this point.