Eliminate the use of LovelaceError in genesis generation.
And improve the error messages at the same time. LovelaceErrors are no
longer generated by the lovelace operations, after previous refactoring
steps they were being thrown locally. Replace the general use of
LovelaceError with a local error that is more specific. The error is
really that one quanity was expected to be bigger than another and it
was not.