Specialise the TxValidationLovelaceError to its actual use
It is now only used in one case, for when the input and output balances
of the transaction do not match (which corresponds with when doing the
lovelace subtraction operation to calculate the fee would fail).

The binary format is retained for compatability. We rename the
constructor to reflect its use. We eliminate the unused field. The
values fields are changed to Lovelace (but encoded as Word64).