Add CompactLovelace type to keep CompactTxOut compact
The CompactTxOut type used to contain simply a Lovelace, which was fine
when it was a newtype wrapper over Word64, but when we made Lovelace a
Natural then it is no longer so small.

So we add and use a CompactLovelace type.

Why is this OK? Because the sum of the whole utxo will fit in a Word64.