Allow software updates to keep the same app version
The existing code comment says that we allow keeping the same version,
but the existing code always required an increment of 1.

A superficial review of the legacy code suggests it also allows either
keeping the same version or an increment of 1.

This is an experimental change. It has to be reviewed against the formal
spec and legacy cardano-sl.