Extend Attributes and Address test generator to include UnparsedFields
The Attributes type can contain UnparsedFields but the generator was
relying on the smart constructors mkAttributes and makeAddress which
always used empty UnparsedFields. We need UnparsedFields for proper test
coverage.

Two tests rely on Attributes with no UnparsedFields: tests that are
checking the size of the Attributes. This is reasonable. So they now use
a new generator variant genAttributesNoUnparsedFields.